[Next Gen] refactor(codegen): split static-element serialization into static_element
